feat: Moving code for syncing user details to home page
